feat(capture): command ring — radial capture-mode picker (CP-0007) (#52)
* feat(capture): command ring — radial capture-mode picker overlay (CP-0007)

Add a global-hotkey radial "command ring" that spawns at the cursor and lets
you pick a capture mode with the mouse. Four wedges — window (top), full
(right), scroll (bottom), area (left) — matching the mockup; clicking a wedge
runs that capture exactly like the tray/hotkey paths and dismisses the ring.
Esc, click-outside, or focus-loss dismiss without capturing.

Interaction model: toggle-open + click (press the hotkey to open; no hold
required). Default hotkey CmdOrCtrl+Shift+Space, editable in Settings ▸
Shortcuts. Gives `scroll` its first hotkey-reachable entry point.

Rust:
- shortcuts: new HotkeyAction::CommandRing + hotkeys.commandRing config key,
  routed to windows::show_command_ring (bypasses capture dispatch like
  ShowEditor).
- windows: show_command_ring builds a transparent, always-on-top `command-ring`
  window, positions it centered on the cursor's monitor via a pure, unit-tested
  ring_position() clamp (never overflows a display; mirrors hud_position).
  Cross-platform cursor: cursor_cg_point on macOS (logical CG), Window::
  cursor_position elsewhere (physical) — matching xcap's per-platform units.
- commands::command_ring: command_ring_select (close ring, then dispatch the
  capture) + close_command_ring.
- overlay capability gains the `command-ring` window label.

Frontend:
- app/ring: SVG radial menu — hover highlight, angle→wedge hit-testing,
  transparent-window handling (scroll-HUD pattern), focus/Esc/blur dismissal.
- lib/commandRing: pure wedge geometry (wedgeAtPoint) + tests.
- config/shortcuts/Settings: commandRing hotkey wiring + recorder row.

Also silence pre-existing clippy lints surfaced by a newer toolchain
(too_many_arguments on hud_position + test helpers; a redundant cast in stitch).


* fix(capture): swallow ring setFocus rejection (no JS set-focus permission)

getCurrentWindow().setFocus() rejects with a permission error (core:window
:default does not grant allow-set-focus, by design). The window is already
focused from Rust in show_command_ring, so the JS call is best-effort — catch
the rejection like the area overlay does instead of letting it surface as a
Next.js runtime error.


* style(capture): make command ring backdrop fully opaque

Solid #17171c disc instead of a 72%-alpha panel so the desktop no longer shows
through the ring; stronger hover fill for clarity on the solid surface. Window
outside the disc stays transparent so the ring keeps its floating-circle shape.


* feat(capture): command ring center editor button + light/dark theming

- Center dead-zone is now an editor button: clicking it opens/refocuses the
  editor (new command_ring_editor: close ring, then windows::show_editor).
  The whole inner circle is the hit-area; wedges stay in the ring band; a
  click beyond the outer edge still dismisses.
- Ring follows the app theme (light/dark). ThemeManager already toggles
  .dark/.light on <html> in every window incl. the ring; colors now use the
  CSS-var palette via `style` (var() does not resolve in SVG presentation
  attributes). Panel uses the opaque --surface-overlay token, so it stays
  100% opaque in both themes.


---------

+++ src-tauri/src/services/stitch.rs
@@ -754,7 +754,7 @@ mod tests {
     fn vnoise_image(w: u32, h: u32) -> RgbaImage {
         let mut img = RgbaImage::new(w, h);
         for y in 0..h {
-            let mut v = (y as u32).wrapping_add(1).wrapping_mul(2_654_435_761);
+            let mut v = y.wrapping_add(1).wrapping_mul(2_654_435_761);
             v ^= v >> 15;
             v = v.wrapping_mul(2_246_822_519);
             v ^= v >> 13;
